configured critic_config file should not be optional
If you configure a critic_config option, it shouldn't be ignored just
because it doesn't exist. Instead, it should cause an error.

Rather than having a default value for critic_config, calculate the
correct file in the munge phase. If not configured, we can default to
'perlcritic.rc' if the file exists. Otherwise, leave Test::Perl::Critic
to follow its default behavior.

This should match the previous behavior, except will cause errors if you
configure a file that doesn't exist.